Slight optimizations to collision, weapons, and procgen code
* Make the `MAP_INDEX` and `MAP_INDEX_ROW_OFFSET` macros more resistant to integer overflow bugs.
* Use pointer arithmetic instead of array accessors as the former is more performant according to the [coding guidelines](https://gbdk-2020.github.io/gbdk-2020/docs/api/docs_coding_guidelines.html).

+++ src/common.h
@@ -43,8 +43,8 @@
  */
 #define TILE_SIZE_BYTES 16
 // Bit-shift by 5 is equivalent to multiply by 32, which is the ROW_WIDTH.
-#define MAP_INDEX(row, column) (((row) << 5) + column)
-#define MAP_INDEX_ROW_OFFSET(row) ((row) << 5)
+#define MAP_INDEX(row, column) ((((uint16_t)(row)) << 5) + (column))
+#define MAP_INDEX_ROW_OFFSET(row) (((uint16_t)(row)) << 5)
 // Note: Indexes 0x01-0x24 are used for font characters.
 #define EMPTY_TILE        0x00
 #define WALL_BLOCK_TILE   0x25
